move errorWriting function. Add name to err payload
